--- server/src/adapters/data/mytext.txt
@@ -1,2 +1,0 @@
-aqui van las clases que interactúan con la base de datos o cualquier otra fuente de datos.
-
