--- server/src/adapters/errors/mytext.txt
@@ -1,1 +1,0 @@
-aqui van las clases de errores personalizados que pueden ser utilizadas en toda la aplicación.
