--- server/src/application/usecases/mytext.txt
@@ -1,1 +1,0 @@
-aqui van los casos de uso de la aplicación.
