--- server/src/config/mytext.txt
@@ -1,1 +1,0 @@
-aqui van los archivos de configuración, como la configuración de la base de datos y la configuración del servidor.
