+++ server/src/controllers/userController.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import {User} from '../../domain/models/User'
+import {User} from '../domain/models/User'
 
 const users: User[] = [];
